Add Ubuntu 14.04 & later support

+++ bin/install.sh
@@ -51,17 +51,53 @@ function os_type
 os_type
 
 # Install Debian Dependencies
-if [[ OS_DEBIAN -gt 0 ]]; then
-    # Test for minimal version
-    VERSION=$(cat /etc/debian_version)
-    if [[ "${VERSION:0:1}" != "8" ]]; then
-        echo "[ERROR] Minimal Debian version of 8 (jessie) required."
-        exit 1
+if [[ OS_DEBIAN -gt 0 ]] || [[ OS_UBUNTU -gt 0 ]]; then
+
+    if [[ OS_DEBIAN -gt 0 ]]; then
+        # Test for minimal version
+        VERSION=$(cat /etc/debian_version)
+        if [[ "${VERSION:0:1}" != "8" ]]; then
+            echo "[ERROR] Minimal Debian version of 8 (jessie) required."
+            exit 1
+        fi
+        echo "[INFO] Installing Debian specific dependencies..."
+        apt-get -qq -y install elasticsearch
+        if [[ $? -ne 0 ]]; then
+            echo "[ERROR] Cannot install Debian specific dependencies."
+            exit
+        fi
+        echo "[OKAY] Debian specific dependencies installed"
     fi
+
+    if [[ OS_UBUNTU -gt 0 ]]; then
+        . /etc/lsb-release
+        if [[ ${DISTRIB_RELEASE:0:2} -lt 14 ]]; then
+            echo "[ERROR] Minimal Ubuntu version of 14.04 (trusty) required."
+            exit 1
+        fi
+
+        echo "[INFO] Installing Ubuntu specific dependencies..."
+
+        # Install Elasticsearch
+        apt-get -qq -y install openjdk-7-jre-headless
+        if [[ $? -ne 0 ]]; then
+            echo "[ERROR] Cannot install Debian specific dependencies."
+            exit
+        fi
+
+        wget --no-check-certificate -O /tmp/elasticsearch-1.7.2.deb \
+            https://download.elastic.co/elasticsearch/elasticsearch/elasticsearch-1.7.2.deb && dpkg -i /tmp/elasticsearch-1.7.2.deb
+        if [[ $? -ne 0 ]]; then
+            echo "[ERROR] Could not install Ubuntu specific elasticsearch."
+            exit
+        fi
+        echo "[OKAY] Ubuntu specific dependencies installed"
+    fi
+
+
     # Apt Dependencies
     echo -n "Installing Apt Dependencies..."
-    apt-get -qq -y install git python3 python3-requests python3-yaml python3-dnspython python3-pip python3-dateutil \
-    elasticsearch
+    apt-get -qq -y install git python3 python3-requests python3-yaml python3-dnspython python3-pip python3-dateutil
     if [[ $? -ne 0 ]]; then
         echo "[ERROR] Cannot Apt Install dependencies."
         exit
@@ -80,8 +116,15 @@ if [[ OS_DEBIAN -gt 0 ]]; then
     # Modify elasticsearch to startup automatically
     cat /etc/default/elasticsearch | sed -e 's/#START_DAEMON/START_DAEMON/' > /etc/default/elasticsearch.new
     mv /etc/default/elasticsearch.new /etc/default/elasticsearch
-    systemctl stop elasticsearch
-    systemctl start elasticsearch
+
+    if [[ OS_DEBIAN -gt 0 ]]; then
+        systemctl stop elasticsearch
+        systemctl start elasticsearch
+    fi
+    if [[ OS_UBUNTU -gt 0 ]]; then
+        service elasticsearch stop
+        service elasticsearch start
+    fi
 
     # Wait for a little bit for elastic search to start up
     ES_STARTED=0
@@ -111,9 +154,18 @@ if [[ OS_DEBIAN -gt 0 ]]; then
     chown cif:cif -Rf /opt/cifpy3
 
     # Copy systemd scripts
-    cp /opt/cifpy3/scripts/debian/cif-server.systemd /etc/systemd/system/cif-server.service
-    cp /opt/cifpy3/scripts/debian/cif-server.default /etc/default/cif-server
-    
+    if [[ OS_DEBIAN -gt 0 ]]; then
+        cp /opt/cifpy3/scripts/debian/cif-server.systemd /etc/systemd/system/cif-server.service
+        cp /opt/cifpy3/scripts/debian/cif-server.default /etc/default/cif-server
+    fi
+    if [[ OS_UBUNTU -gt 0 ]]; then
+        cp /opt/cifpy3/scripts/ubuntu/cif-server.upstart /etc/init/cif-server.conf
+        cp /opt/cifpy3/scripts/ubuntu/cif-server.default /etc/default/cif-server
+    fi
+
+    # Download GeoIP data
+    /opt/cifpy3/bin/cif-utility -g
+
     # Run the cif initial install
     TOKEN=$(/opt/cifpy3/bin/cif-utility -r)
     
@@ -125,9 +177,13 @@ if [[ OS_DEBIAN -gt 0 ]]; then
     . /etc/profile.d/cif.sh
 
     # Start it up, need to detect which version
-    systemctl enable cif-server
-    systemctl start cif-server
-    
+    if [[ OS_DEBIAN -gt 0 ]]; then
+        systemctl enable cif-server
+        systemctl start cif-server
+    fi
+    if [[ OS_UBUNTU -gt 0 ]]; then
+        service cif-server start
+    fi
     # Print information
     echo "[OKAY] CIF has been installed. You can now use the 'cif' command."
     echo "[OKAY] Your ADMIN API token is ${TOKEN}. It has also been written to ~/.cif"
